zonec.h: (zparser_type) Changed type of 'line' member.
zonec.c: (error, warning) Use correct format for printing line.


git-svn-id: file:///svn/nsd/trunk@972 a26ef69c-88ff-0310-839f-98b793d9c207

+++ zonec.h
@@ -78,7 +78,7 @@ struct zparser {
 	domain_type *prev_dname;
 	unsigned int _rc;   /* current rdata cnt */
 	unsigned int errors;
-	size_t line;
+	unsigned int line;
 	const char *filename;
 };
